Add VendorWait activity screen.
Still need to actually show the vendors nearby.

Handle the intent between hitting the 'X' mark and going back to the previous Activity.  In the process,
we cancel the intent.

Create a bundles directory for all the Intent bundle data we pass back.  Intents contain Bundles of data, so
I figure it's better to call it this name.